abstract game
added spanish and american versions
